Create email_metrics_per_account_by_month.sql
This SQL query calculates the percentage of emails sent to each account within a given month by comparing the number of emails sent to the account with the total number of emails sent across all accounts in that month. It also determines the date of the first and last email sent to each account within the month, using window functions for data calculation and aggregation.